Vilesci: - Personensuche jetzt auch nach Alias möglich - LV-Wartung CopyPaste Fehler in Überschrift behoben - Gruppenverwaltung: Personen in Generierten Gruppen können nicht mehr hinzugefügt/entfernt werden - Institutsliste: ALVS werden jetzt richtig sortiert - Tippfehler im Menü behoben - Personenzusammenlegung: Suche ist jetzt ein Feld und caseinsensitiv - Erweiterung der Preinteressentenverwaltung

+++ vilesci/lehre/einheit_det.php
@@ -6,6 +6,7 @@ require_once('../../include/benutzergruppe.class.php');
 require_once('../../include/person.class.php');
 require_once('../../include/benutzer.class.php');
 require_once('../../include/student.class.php');
+require_once('../../include/gruppe.class.php');
 
 if(!$conn=pg_pconnect(CONN_STRING))
    die('Fehler beim Aufbau der Datenbankconnection');
@@ -30,6 +31,9 @@ else if (isset($_GET['type']) && $_GET['type']=='delete')
 	$e=new benutzergruppe($conn);
 	$e->delete($_GET['uid'], $kurzbz);	
 }
+$gruppe = new gruppe($conn);
+if(!$gruppe->load($kurzbz))
+	die('Gruppe wurde nicht gefunden');
 
 ?>
 <html>
@@ -40,29 +44,32 @@ else if (isset($_GET['type']) && $_GET['type']=='delete')
 </head>
 <body>
 <H1>Gruppe <?php echo $kurzbz ?></H1>
-
-<FORM name="newpers" method="post" action="einheit_det.php">
-  <INPUT type="hidden" name="type" value="new">
-    
-  <SELECT name="uid">
-    <?php
-    	$qry = "SELECT * FROM campus.vw_benutzer ORDER BY nachname, vorname";
-    	echo $qry;
-    	$result = pg_query($conn, $qry);
-    			
-		for ($i=0;$row = pg_fetch_object($result);$i++)
-		{			
-			echo "<option value=\"".$row->uid."\">".$row->nachname." ".$row->vorname." - ".$row->uid."</option>";
-		}
-		?>
-  </SELECT>
-  
-  <INPUT type="hidden" name="einheit_id" value="<?php echo $kurzbz; ?>">
-  <INPUT type="submit" name="new" value="Hinzuf&uuml;gen">
-</FORM>
-<HR>
-
 <?php
+if(!$gruppe->generiert)
+{
+	echo '
+	<FORM name="newpers" method="post" action="einheit_det.php">
+	  <INPUT type="hidden" name="type" value="new">
+    
+  	<SELECT name="uid">';
+   
+	$qry = "SELECT * FROM campus.vw_benutzer ORDER BY nachname, vorname";
+	echo $qry;
+	$result = pg_query($conn, $qry);
+			
+	for ($i=0;$row = pg_fetch_object($result);$i++)
+	{			
+		echo "<option value=\"".$row->uid."\">".$row->nachname." ".$row->vorname." - ".$row->uid."</option>";
+	}
+
+	echo '
+	  </SELECT>
+	  
+	  <INPUT type="hidden" name="einheit_id" value="<?php echo $kurzbz; ?>">
+	  <INPUT type="submit" name="new" value="Hinzuf&uuml;gen">
+	</FORM>
+	<HR>';
+}
 	$qry = "SELECT * FROM public.tbl_benutzergruppe, public.tbl_benutzer, public.tbl_person WHERE".
 	       " tbl_benutzergruppe.gruppe_kurzbz='".addslashes($kurzbz)."' AND".
 	       " tbl_benutzergruppe.uid = tbl_benutzer.uid AND tbl_benutzer.person_id=tbl_person.person_id ORDER BY nachname, vorname";
@@ -80,7 +87,8 @@ else if (isset($_GET['type']) && $_GET['type']=='delete')
 		    echo "<td>".$row->uid."</td>";
 			echo "<td>".$row->vorname."</td>";
 			echo "<td>".$row->nachname."</td>";
-			echo '<td class="button"><a href="einheit_det.php?uid='.$row->uid.'&type=delete&kurzbz='.$kurzbz.'">Delete</a></td>';
+			if(!$gruppe->generiert)
+				echo '<td class="button"><a href="einheit_det.php?uid='.$row->uid.'&type=delete&kurzbz='.$kurzbz.'">Delete</a></td>';
 		    echo "</tr>\n";
 		}
 	}
